Fix problem with single member groups
If we request members from a group with only one member, we get a single
user hash instead of an array of hashes. Trying to iterate the hash was
causing the process to break. Solution is to wrap the hash in an array so
we can iterate it as normal.